npm: move native helpers to npm6 namespace
Moving npm native helpers to npm6 namespace in preparation for adding
npm7 support.

The pattern is slightly different from composer because of the way we
invoke the helper with a `package_manager:function` rpc syntax.
