+++ npm_and_yarn/helpers/lib/npm6/conflicting-dependency-parser.js
