+++ npm_and_yarn/helpers/lib/npm6/remove-dependencies-from-lockfile.js
