+++ npm_and_yarn/helpers/test/npm6/fixtures/conflicting-dependency-parser/deeply-nested/package.json
