+++ npm_and_yarn/helpers/test/npm6/fixtures/conflicting-dependency-parser/nested/package.json
