+++ npm_and_yarn/helpers/test/npm6/fixtures/conflicting-dependency-parser/simple/package.json
